fix(ci): split terminal module for cross-platform + gate unix-only tests
- Split cli/terminal/mod.rs into unix.rs (existing code) + windows.rs
  (minimal StreamManager stub with Ctrl+C via tokio::signal::ctrl_c)
- Gate integration tests that use Unix-specific APIs with #![cfg(unix)]:
  recovery.rs (libc::kill), copy.rs (unix::fs::symlink),
  mount_security.rs (unix::fs::MetadataExt),
  sigstop_quiesce.rs (libc::kill/SIGSTOP)

The Windows CI workflow compiles all targets (--all-targets) even though
it only runs unit tests, so test code must also compile on Windows.
